added person registration

+++ src/main/java/com/agat/test/repos/UserRepo.java
@@ -3,6 +3,6 @@ package com.agat.test.repos;
 import com.agat.test.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<User, Long> {
+public interface UserRepo extends JpaRepository<User, Integer> {
     User findByUsername(String username);
 }
